extract and document TempConfigurationGitRepo
